remove procedure package
The idea behind the procedure package was to hold storage
operation sequences that needed to be shared between multiple
components. However, it has limited use and several downsides:

- When the underlying operations are stored in a separate package, they
  must be public, even if they are only used in a procedure. This
  unecessarily increases the surface area for misuse of these functions.
- Not all storage operation sequences are actually kept in the procedure
  pacakge. In practice, there are very few older functions here, while new
  "procedures" are kept in "operation" or in higher-level storage
  modules.
