+++ state/cluster/badger/snapshot_test.go
@@ -20,7 +20,6 @@ import (
 	"github.com/onflow/flow-go/storage"
 	"github.com/onflow/flow-go/storage/operation"
 	"github.com/onflow/flow-go/storage/operation/pebbleimpl"
-	"github.com/onflow/flow-go/storage/procedure"
 	"github.com/onflow/flow-go/storage/store"
 	"github.com/onflow/flow-go/utils/unittest"
 )
@@ -149,7 +148,7 @@ func (suite *SnapshotSuite) Proposal() model.Proposal {
 func (suite *SnapshotSuite) InsertBlock(proposal model.Proposal) {
 	err := unittest.WithLock(suite.T(), suite.lockManager, storage.LockInsertOrFinalizeClusterBlock, func(lctx lockctx.Context) error {
 		return suite.db.WithReaderBatchWriter(func(rw storage.ReaderBatchWriter) error {
-			return procedure.InsertClusterBlock(lctx, rw, &proposal)
+			return operation.InsertClusterBlock(lctx, rw, &proposal)
 		})
 	})
 	suite.Require().NoError(err)
@@ -239,7 +238,7 @@ func (suite *SnapshotSuite) TestFinalizedBlock() {
 	// finalize the block
 	err = unittest.WithLock(suite.T(), suite.lockManager, storage.LockInsertOrFinalizeClusterBlock, func(lctx lockctx.Context) error {
 		return suite.db.WithReaderBatchWriter(func(rw storage.ReaderBatchWriter) error {
-			return procedure.FinalizeClusterBlock(lctx, rw, finalizedProposal1.Block.ID())
+			return operation.FinalizeClusterBlock(lctx, rw, finalizedProposal1.Block.ID())
 		})
 	})
 	suite.Require().NoError(err)
