+++ storage/store/cluster_blocks_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/onflow/flow-go/storage"
 	"github.com/onflow/flow-go/storage/operation"
 	"github.com/onflow/flow-go/storage/operation/dbtest"
-	"github.com/onflow/flow-go/storage/procedure"
 	"github.com/onflow/flow-go/utils/unittest"
 )
 
@@ -39,7 +38,7 @@ func TestClusterBlocks(t *testing.T) {
 			// InsertClusterBlock only needs LockInsertOrFinalizeClusterBlock
 			err := unittest.WithLock(t, lockManager, storage.LockInsertOrFinalizeClusterBlock, func(lctx2 lockctx.Context) error {
 				return db.WithReaderBatchWriter(func(rw storage.ReaderBatchWriter) error {
-					return procedure.InsertClusterBlock(lctx2, rw, unittest.ClusterProposalFromBlock(block))
+					return operation.InsertClusterBlock(lctx2, rw, unittest.ClusterProposalFromBlock(block))
 				})
 			})
 			require.NoError(t, err)
@@ -47,7 +46,7 @@ func TestClusterBlocks(t *testing.T) {
 			// FinalizeClusterBlock only needs LockInsertOrFinalizeClusterBlock
 			err = unittest.WithLock(t, lockManager, storage.LockInsertOrFinalizeClusterBlock, func(lctx2 lockctx.Context) error {
 				return db.WithReaderBatchWriter(func(rw storage.ReaderBatchWriter) error {
-					return procedure.FinalizeClusterBlock(lctx2, rw, block.ID())
+					return operation.FinalizeClusterBlock(lctx2, rw, block.ID())
 				})
 			})
 			require.NoError(t, err)
